[Refactor] Remove redundant int() cast around round() in cli_translator.py
In Python 3, round() returns an integer when the ndigits argument is omitted or None.
Therefore, casting the result of round() to int() is redundant.

Test: atest atest_unittests
Flag: TEST_ONLY
OFF_PEAK_PRESUBMIT: true

+++ atest/cli_translator.py
@@ -133,7 +133,7 @@ class CLITranslator:
     self._indexing_thread.join()
     metrics.LocalDetectEvent(
         detect_type=DetectType.WAIT_FOR_INDEXING_MS,
-        result=int(round((time.time() - start_wait_for_indexing) * 1000)),
+        result=round((time.time() - start_wait_for_indexing) * 1000),
     )
 
   # pylint: disable=too-many-locals
